requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
dataclasses-json==0.5.4
Django==3.2.4
django-redis==5.0.0
djangorestframework==3.12.4
idna==2.10
iniconfig==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-deadfixtures==2.2.1
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-pythonpath==0.7.3
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-mock==1.9.3
six==1.16.0
sqlparse==0.4.1
stringcase==1.2.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
